Informational tooltips added on hover, and bug fixes
Added support for tooltips by hovering over the Information Source
symbol next to the label for each part of the calculator.
Added "total frames of hitstun" to the output.

Bug fix: "pixels per second"  --> "pixels per frame"
This was a mistake I made in the output. Pixels per second would be VERY
slow!